colision

+++ Block.cpp
@@ -4,7 +4,7 @@
 
 
 Block::Block(LPCWSTR bitmapPath, Graphics *gfx, char symbol, 
-				bool collisionEnabled, short index, short maxGeneratingHeight, short seed) {
+				bool collisionEnabled, int index, int maxGeneratingHeight, int seed) {
 	this->symbol = symbol;
 	this->collisionEnabled = collisionEnabled;
 	this->gfx = gfx;
@@ -20,7 +20,7 @@ Block::~Block() {
 
 
 
-void Block::Render(double x, double y, double offset){
+void Block::Render(int x, int y, int offset){
 	sprites->Draw(index, x*spriteWidth + offset, y*spriteHeight);
 }
 
@@ -28,13 +28,13 @@ void Block::Render(double x, double y, double offset){
 
 
 
-short Block::GetSeed() {
+int Block::GetSeed() {
 	return seed;
 }
-short Block::GetMaxGenerationHeight() {
+int Block::GetMaxGenerationHeight() {
 	return maxGeneratingHeight;
 }
-short Block::GetIndex() {
+int Block::GetIndex() {
 	return index;
 }
 char Block::GetSymbol() {
